@@ -15,10 +15,11 @@
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@astrojs/svelte": "^5.0.3",
+    "@astrojs/svelte": "^7.1.0",
     "@astrojs/tailwind": "^5.1.0",
-    "astro": "^4.0.9",
-    "svelte": "^4.2.5",
+    "@sveltejs/vite-plugin-svelte": "^5.1.0",
+    "astro": "^5.9.1",
+    "svelte": "^5.33.17",
     "tailwindcss": "^3.4.0"
   },
   "devDependencies": {
